Change stepwise sorting algorithm approach
Relying on the built-in sorting algorithm results in much shorter and
cleaner code since it saves us from having to implement an inside-out
version of a sorting algorithm ourselves, but it's at the cost of extra
comparisons. We could get potentially get the best of both worlds if we
used a package that implements bottom-up heapsort.